marginalize function added
marginalize takes an expr and a random variable and marginalizes that random variable out of the expr, either through a sum or an integral

Computing expectation using marginalize now rather than PDF

Also subclassed Events to AtomicEvents. These are iterated out and have value and dict properties

added is_bounded properties to probability spaces and random variables

Added check in core/function/count_ops to not continue traversing a tree once you hit a symbol. This was troublesome because RandomVariables are symbols which have args which should not be traversed in this case.