PDF split to PDF and PMF - Cont distrs added
PDF used to be used for both continuous probability density functions and discrete probability mass functions. There are now two separate functions for this. The mixed case is considered to be a probability density function.

Also, added in a few continuous probability distributions to better test the limits of intgration.

ProbabilityMeasures now have a sample space attribute so they know not to integrate beyond what is defined.