Implement some differentiation of G-functions.
This commit adds differentiation of G-functions with respect to a parameter,
where the definining integrand consists of only one gamma function,
corresponding to this parameter, all other gammas having cancelled to produce
rational functions.

In this case the derivative is again expressible as a sum of G functions.

If you know of more cases in which the derivative with respect to parameters
can be expressed using G-functions, please let us know!