Fix a `bug' in evalf regarding the subs argument.
When functions where changed to automatically evalf(), the core
evalf() code should have been updated (since automatic evalf breaks
precision tracking).

This commit adds a function evalf_subs() to core/evalf.py that
updates the precision of the floats in subs to the desired prec.
Using this in the correct locations makes precsion tracking work
again. This commit also adds a test for this behaviour.

Finally, fix a bug in evalf_subs that, which incorrectly modified
`options` in-place. Add a test for that as well.