Implement hyperexpand() at special points.
Both for meijer g-functions and hypergeometric functions it can happen that
while we may not be able to evaluate F(z) in closed form for general z, we are
actually only interested in F(1) [say], and this *does* happen to have a closed
form expression.

Add support for some such cases.

For g-functions, add a specific test which allows us to extend (generalised)
slater expansions, valid for |z|<1 and/or |z|>1 to z=1, the test guarantees
that both expansions agree there, and with the true value of G(1), even if
G represents two distinct holomorphic functions (i.e. this test ensures
continuity on positive reals, even if G is not holomorphic in any
neighbourhood of 1).

For hypergeometric functions, implement a function
hyperexpand_special(ap, bq, 1) which is automatically called by hyperexpand()
[actually by _hyperexpand and _meijergexpand] "after-the-fact".
hyperexpand_special() tries to recognise forms where special summation
theorems apply. Currently implemented are Gauss' and Kummer's theorems.

hyperexpand_special() is very ad hoc, and also very incomplete. There is
vast literature on this subject, but I am not very familiar with it. For
now it seems best to extend hyperexpand_special() as need arises, and only
if this function becomes disproportionally long to investigate other
algorithms.