Replace Mod1 by Mod1Effective.
The Mod1 object is used to decide if two expressions agree up to integer
differences. It used to be implemented as simplify(a-b).is_integer.

However, in all cases where it is used, it is actually necessary that a and b
differ by an Integer, i.e. a concrete number like 7.

Rename Mod1 to Mod1Effective, change the implementation to
simplify(a-b).is_Integer, change all references to Mod1, and add a test
that used to hang with the old implementation.