Improve performance of combsimp.
Implement a heuristic for absorbing factors into gamma functions:

e.g. x*gamma(y) can be written as C*gamma(y+1), and if C is simpler
than x, this is desirable. Doing this on many factors is slow, so we implement
a heuristic that tries to know in advance which pairs of x, y will not work.
See the code for a description.

Also add a new benchmark for this.